api-login/requirements.txt
fastapi==0.100.0
firebase-admin==6.0.1
python-dotenv==1.0.0
uvicorn==0.22.0
requests==2.28.1
sqlalchemy[asyncio]==2.0.0
asyncpg==0.27.0
